ft-#11: add new property to childcare mongoose schema and update test profile payload
