+++ test/routes/integration/registerChild.test.ts
@@ -6,6 +6,7 @@ import signInUser from "./test-utils/signin";
 import _ from "lodash";
 import userPayload from "./test-utils/signupPayload";
 import axios from "axios";
+import profile_payload from "./test-utils/profilePayload";
 
 let axiosMock = jest.mock("axios")
 axios.get = jest.fn().mockResolvedValue({data: [{"latitude": 1.0, "longitude": 2.1}]})
@@ -15,18 +16,7 @@ describe("Sends request to /create-article", () => {
   let daycareId: string; 
   let userId: string
 
-let profile = {
-    title:  "David's Daycare",
-    amount: "50",
-    from: "2024-10-20",
-    to: "2024-12-12",
-    rating: 5,
-    description: "Am gonna do you well",
-    phonenumber: "0099999999",
-    isOpen: "yes",
-    image: "daycare.png",
-    location: "Abuja,Lagos"
-}
+
 
   afterAll(async () => {
     await mongoose.connection.dropDatabase();
@@ -38,7 +28,7 @@ let profile = {
  await signupUser(userPayload);
     let res = await signInUser(_.pick(userPayload, ["email", "password"]));
     token = res.header.authorization;   
-    let response = await request(app).post("/create-childcare-profile").send(profile).set("authorization", token)
+    let response = await request(app).post("/create-childcare-profile").send(profile_payload).set("authorization", token)
     daycareId = response.body._id
     userId = res.body.message._id
   });
