+++ test/routes/integration/test-utils/profilePayload.ts
@@ -1,0 +1,17 @@
+let profile_payload = {
+    title:  "David's Daycare",
+    amount: "50",
+    from: "2024-10-20",
+    to: "2024-12-12",
+    rating: 5,
+    description: "Am gonna do you well",
+    phonenumber: "0099999999", 
+    isOpen: "yes",
+    image: "daycare.png",
+    location: "Abuja,Nigeria",
+    userId: "659bdb0ad66c81e2ac3e5628",
+    owner: "Peter Parker",
+    role: "Teacher"
+}
+
+export default profile_payload
